pyproject.toml
[project]
name = "retail-insights-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.4",
    "duckdb>=1.4.4",
    "google-genai>=1.66.0",
    "grandalf>=0.8",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "langsmith>=0.7.14",
    "openai>=2.26.0",
    "openpyxl>=3.1.0",
    "pandas>=2.3.3",
    "pymupdf>=1.25.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
